Only set weights returned by Transform.
Currently, ModelTransformer expects all the weights
required by a layer to be returned in the Transform
and set. This has issues since the Transform might
not know all the weights, or might only want to set
some of the weights, not all. Some of the weights
are often created by the layer in the build method.

Example: ConvBatchNorm layers create quant variables
which are determined by the weights and quantizers.
The Transform does not know about them.

This CL allows the Transform to set specific weights
it wants. Sets only the map values, not the entire list.

PiperOrigin-RevId: 278967295