Model clustering for Model Optimization Toolkit implementation
This PR introduces a model clustering technique for Model Optimization
Toolkit. In this context the word "clustering" means when model
weights have only a few unique values. This allows for certain types
of hardware to benefit from advanced weight compression techniques
and the associated reduction in model memory footprint and bandwidth.
Clustering can also be combined with pruning and quantization reducing
network’s storage requirements even further.

The technique description that is implemented can be found here:
https://arxiv.org/abs/1510.00149

From practical standpoint this is implemented using a lookup table to
hold the cluster centroid values during model training. The weight
array is populated with 'gather' operation so that during back
propagation the gradients can be calculated in a normal way. The lookup
table is then adjusted using the cumulative gradient values for the
weights that correspond to the same centroid.

The API is intended to be used in the same way pruning API is used.
The interfaces is intentionally made very similar to the one described
here:
https://www.tensorflow.org/model_optimization/guide/pruning/

The usage of this API can be as simple as this:

  ```python
  clustering_params = {
    'number_of_clusters': 8,
    'cluster_centroids_init': 'density-based'
  }

  clustered_model = cluster_weights(original_model, **clustering_params)
  ```
An example of the API in use can be found at python/example/clustering

*This PR does not affect any of the existing pruning functionality and
is completely independent from the rest of the existing code base.*

* cluster.py/cluster_test.py - high level clustering interfaces and tests
* cluster_wrapper.py/cluster_wrapper_test.py - an implementation of a
particular clustering algorithm and associated tests
clusterable_level.py - similar to PrunableLayer, provides a wrapper for
a layer to be recognized as a clusterable one.
* clustering_centroids.py/clustering_centroids_test.py - implementations
of different clusters centroids initialization mechanics and tests for
them
* clustering_registry.py/clustering_registry_test.py - similar to
PruningRegistry contains the list of clusterable parameters in each
layer.