Create NoOpActivation to not affect quantization
Current quantization strategy applies a FakeQuant after
`linear` activations. We need support for cases where we
don't quantize such no-op activations. Adding a new NoOpActivation
specific to quant code so this can be distinguished.

Example, for Conv->BatchNorm, when not folding the
BatchNorm in we need to ensure no Quant operation
happens between the layers.

PiperOrigin-RevId: 286455419