Store batch mean/variance as private fields in BatchNorm
Quantization code performs folding operations on BatchNorm
layers which requires access to the batch mean and variance
calculated during training. The quantization code then
accesses these moments for further calculation.

Currently, the BatchNorm layer does not expose these tensors.
Storing them locally in a private field for now so it doesn't
affect the interface of BatchNorm, and allows the quantization
code to function.

PiperOrigin-RevId: 286497355