Working version of new abstractions for model optimziation algorithms.
With the proposed design, `OptimizationAlgorithmInterface` and `StateRepr` would be the general symbols exposed to researchers and contributors, wishing to implement a novel model optimization algorithm.

Exposed symbols specific to Keras would be `wrap_keras_layer` method and `KerasOptimizationAlgorithmConfig`.

As part of private API would be the `ModelOptimizationWrapper`, which takes care of all Keras-specific integration.

An example implementation of existing pruning technique in terms of this proposed API is provided in pruning_algorithm.py.

Limitations of the current snapshot:
- Only graph execution is supported.
- Naming and documentation is not fully settled, and likely to change.
- Implementation of `ModelOptimizationWrapper` is "hacky" and slightly inefficient - and needs to be improved. As being part of private API, it can be iterated on without major breakage concerns.

PiperOrigin-RevId: 288598196