Fix infinite loop when pattern can match continuously
Earlier, we were preventing inifinite match loops in cases
where the same pattern+replace would continue to match.
However, there are cases where the match+replace might not
be the same, yet still continue indefinitely.

Example: A transform such as A => A->B, will continue to
match A indefinitely.

This CL introduces a map of patterns to specific layers and
ensures this match only happens once.

This is helpful to ensure an InputLayer => InputLayer -> Quantizer
match replaces all the InputLayers in the graph, but only once.

Also adds test to verify.

PiperOrigin-RevId: 293929574