Feature request template.
The template emphasizes motivation to help with prioritizing and demonstrating the usefulness of a change and seeks to be as lean as possible while addressing those needs.

This can be applied to nearly all smaller API changes where a RFC is too much, if usability improvements that require API changes also count as feature requests.

PiperOrigin-RevId: 313859237