Use keras `relu` in guide rather than tf directly
The current example uses `tf.nn.relu` which internally
gets treated as a callable. The quantization machinery
is not expected to know about tf internal functions, just
handle Keras constructs.

This happens to work since the function name matches `relu`.
But we should be using the Keras buit-in activations, which
is what we support.

PiperOrigin-RevId: 314440464