Support for SeparableConv2D for QAT
SeparableConv2D is a combination of DepthwiseConv2D
and Conv2D. QAT needs to break the layer up into this
combination so that the rest of the infra can then apply
QAT to the resulting layers.

Not possible to implement support for SeparableConv directly
since it does not provide hooks to alter internal graph
construction.

PiperOrigin-RevId: 321030603