Add block sparsity pruning in TF MOT.
This CL adds a new option, block_pruning_method, in TF MOT. This option is enabled when user set a block_size greater than (1, 1).

When this option is set to 'INTER', pruning is done at the block level, i.e. one entire block is either pruned or not pruned. This is the block pruning implementation in current TF MOT.

When this option is set to 'INTRA', pruning within each block is performed, which is a new feature. This feature is generalized to support k in block (m, n) sparsity, where m, n, k are positive integers.

PiperOrigin-RevId: 388864074