Bump up tensorflow version to 2.9.1 for continuous build tests.
Previous version 2.7.0 occasionally fails to build because of the version incompatibility between protobuf and tensorboard.
The following is the build error message.

```
ERROR: tensorboard 2.9.1 has requirement protobuf<3.20,>=3.9.2, but you'll have protobuf 4.21.2 which is incompatible.
```

With tensorflow version >=2.7.3, the problem is fixed by the explicit protobuf version range specification: `protobuf<3.20,>=3.9.2`.

However, the unit tests rely on a feature that was introduced in `tensorflow-compression>=2.9.0`, which transitively requires `tensorflow>=2.9.0`.
Simply bumping up tensorflow's version to >=2.7.3 also causes install failures for `tensorflow-compression`.

Therefore tensorflow's version is bumped up to `2.9.1` and the version for `tensorflow-compression` is explicitly required to be compatible with `2.9.1`.

PiperOrigin-RevId: 462517238